+++ WordPress-VIP/ruleset.xml
@@ -93,4 +93,7 @@
 		<message>Usage of a direct database call without caching is prohibited on the WP VIP platform. Use wp_cache_get / wp_cache_set or wp_cache_delete.</message>
 	</rule>
 
+	<!-- https://vip.wordpress.com/documentation/vip/code-review-what-we-look-for/#uncached-pageload -->
+	<rule ref="WordPress.DB.SlowDBQuery"/>
+
 </ruleset>
