+++ WordPress/Tests/DB/SlowDBQueryUnitTest.php
@@ -7,7 +7,7 @@
  * @license https://opensource.org/licenses/MIT MIT
  */
 
-namespace WordPress\Tests\VIP;
+namespace WordPress\Tests\DB;
 
 use PHP_CodeSniffer\Tests\Standards\AbstractSniffUnitTest;
 
@@ -18,9 +18,7 @@ use PHP_CodeSniffer\Tests\Standards\AbstractSniffUnitTest;
  *
  * @since   0.3.0
  * @since   0.13.0 Class name changed: this class is now namespaced.
- * @since   0.15.0 The sniff has been deprecated. This unit test file now
- *                 only tests that the deprecation warnings are correctly thrown
- *                 and that the sniff falls through to the parent correctly.
+ * @since   0.15.0 This sniff has been moved from the `VIP` category to the `DB` category.
  */
 class SlowDBQueryUnitTest extends AbstractSniffUnitTest {
 
@@ -41,8 +39,13 @@ class SlowDBQueryUnitTest extends AbstractSniffUnitTest {
 	 */
 	public function getWarningList() {
 		return array(
-			1 => 2,
-			4 => 1,
+			4  => 1,
+			10 => 1,
+			15 => 1,
+			16 => 1,
+			19 => 2,
+			30 => 1,
+			32 => 1, // Warning about deprecated whitelist comment.
 		);
 
 	}
